Restructure Native Directory
