+++ src/modules/NetworkLogger.ts
@@ -2,7 +2,8 @@ import { Platform } from 'react-native';
 
 import type { RequestHandler } from '@apollo/client';
 
-import { NativeAPM, NativeInstabug } from '../native';
+import { NativeAPM } from '../native/NativeAPM';
+import { NativeInstabug } from '../native/NativeInstabug';
 import IBGEventEmitter from '../utils/IBGEventEmitter';
 import InstabugConstants from '../utils/InstabugConstants';
 import xhr, { NetworkData, ProgressCallback } from '../utils/XhrNetworkInterceptor';
