+++ src/modules/Replies.ts
@@ -1,6 +1,6 @@
 import { Platform } from 'react-native';
 
-import { NativeReplies } from '../native';
+import { NativeReplies } from '../native/NativeReplies';
 import IBGEventEmitter from '../utils/IBGEventEmitter';
 import InstabugConstants from '../utils/InstabugConstants';
