+++ src/native/NativeCrashReporting.ts
@@ -1,6 +1,8 @@
 import type { NativeModule, Platform } from 'react-native';
 import type { StackFrame } from 'react-native/Libraries/Core/Devtools/parseErrorStack';
 
+import { NativeModules } from './NativePackage';
+
 export interface CrashData {
   message: string;
   e_message: string;
@@ -15,3 +17,5 @@ export interface CrashReportingNativeModule extends NativeModule {
   sendJSCrash(data: CrashData | string): void;
   sendHandledJSCrash(data: CrashData | string): void;
 }
+
+export const NativeCrashReporting = NativeModules.IBGCrashReporting;
