+++ src/native/NativeFeatureRequests.ts
@@ -2,6 +2,7 @@ import type { NativeModule } from 'react-native';
 
 import type { actionTypes } from '../utils/ArgsRegistry';
 import type { ActionType } from '../utils/Enums';
+import { NativeModules } from './NativePackage';
 
 export interface FeatureRequestsNativeModule extends NativeModule {
   setEnabled(isEnabled: boolean): void;
@@ -11,3 +12,5 @@ export interface FeatureRequestsNativeModule extends NativeModule {
     types: actionTypes[] | ActionType[],
   ): void;
 }
+
+export const NativeFeatureRequests = NativeModules.IBGFeatureRequests;
